refactor: remove dead code and the vestigial gain_group field
- Delete src/lib/splits/normalization.ts (verbatim duplicate of
  SplitEnricher, used only by its own tests, already drifting);
  retarget its tests at SplitEnricher
- Remove gain_group from EnrichedTransaction: every producer hardcoded
  'NONE' and nothing read it — the CGT engine populates match_groups
- Remove dead store actions addTransactions (which fabricated
  tax_year/fx_rate values) and setHelpContext
- Remove unused parseCSVText and the unused FXManager singleton
  (getDefaultFXManager/resetDefaultFXManager)
- Reuse parsingUtils.parseISODate / extractCurrencyFromAmount in the
  freetrade and revolut parsers instead of local reimplementations
