+++ src/lib/cgt/__tests__/shortSellMatcher.test.ts
@@ -24,7 +24,6 @@ function createTransaction(
     fx_source: 'HMRC',
     fx_error: null,
     tax_year: '2023/24',
-    gain_group: 'NONE',
     ...overrides,
   }
 }
